jkjkjb

+++ Entities/centro_educativo.cs
@@ -10,15 +10,15 @@ namespace practiquesIEI.Entities
     {
         public string nombre { get; set; }
         public tipo_centro tipo { get; set; }
-        public double longitud { get; set; }
-        public double latitud { get; set; }
+        public decimal longitud { get; set; }
+        public decimal latitud { get; set; }
         public int telefono { get; set; }
         public string descripcion { get; set; }
         public string direccion { get; set; }
         public int cod_postal { get; set; }
 
         public centro_educativo() { }
-        public centro_educativo(string nombre, tipo_centro tipo, double longitud, double latitud, int telefono, string descripcion, string direccion, int cod_postal)
+        public centro_educativo(string nombre, tipo_centro tipo, decimal longitud, decimal latitud, int telefono, string descripcion, string direccion, int cod_postal)
         {
             this.nombre = nombre;
             this.tipo = tipo;
